# lib/api-zod/src/generated/types/bedUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BedUpdateStatus } from "./bedUpdateStatus";

export interface BedUpdate {
  propertyId?: string;
  bedNumber?: number;
  room?: string;
  status?: BedUpdateStatus;
  /** @nullable */
  occupantId?: string | null;
}