# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyPaymentMethod } from "./propertyPaymentMethod";
import type { PropertyStatus } from "./propertyStatus";

export interface Property {
  id: string;
  name: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  totalBeds: number;
  monthlyRent: number;
  chargePerBed: number;
  status: PropertyStatus;
  landlordName: string;
  landlordEmail: string;
  landlordPhone: string;
  paymentMethod: PropertyPaymentMethod;
  paymentRecipient: string;
  paymentDueDay: number;
  paymentNotes: string;
  bankName: string;
  bankRouting: string;
  bankAccount: string;
  portalUrl: string;
  notes: string;
  furnishings: string[];
}