# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdatePaymentMethod } from "./propertyUpdatePaymentMethod";
import type { PropertyUpdateStatus } from "./propertyUpdateStatus";

export interface PropertyUpdate {
  name?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  totalBeds?: number;
  monthlyRent?: number;
  chargePerBed?: number;
  status?: PropertyUpdateStatus;
  landlordName?: string;
  landlordEmail?: string;
  landlordPhone?: string;
  paymentMethod?: PropertyUpdatePaymentMethod;
  paymentRecipient?: string;
  paymentDueDay?: number;
  paymentNotes?: string;
  bankName?: string;
  bankRouting?: string;
  bankAccount?: string;
  portalUrl?: string;
  notes?: string;
  furnishings?: string[];
}